==> notebooks/PIPELINE_1_WASTE_COLLECTION_DETECTION/01_FAST_MOVEMENT_MODELING.ipynb <==
# About

* **Author**: Adil Rashitov
* **Creation Date**: 19.02.2020
* **Goal**: This notebook is created for experiments with modeling of vehicle fast movement for clustering needs
* **Deliverable**: Model detecting if car moving fast or not

# Imports
import os
import numpy as np
import pandas as pd
import logging
import plotly.express as px
import geopandas as gpd
import plotly.express as px
from multiprocessing import Pool
from sklearn.model_selection import train_test_split
from GPSOdyssey import Polaris, Kepler, Void, Vega
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = 'all'

%load_ext autoreload
%autoreload 2

# Data

HISTORY = []

## Train GPS records 

I_OUTPUT_GPS_METRICS_ANALYSIS = '/Data/Intermediate/FastMovementDetection/TrainTestFastMovementDetection.csv'
gps = pd.read_csv(I_OUTPUT_GPS_METRICS_ANALYSIS, parse_dates=['datetime'])

gps['trip_id'] = gps['vehicle_id'] + ' ' + gps['datetime'].dt.date.astype('str')

gps.head()

## Deliverable model

OUTPUT_MODEL = '/models/fast_movement_detector.sav'
OUTPUT_MODEL

!ls /models

# Preprocessing

## Feature engineering

unique_trips = gps['trip_id'].unique()
gps_trips = list(map(lambda gps, trip: gps[gps['trip_id'] == trip], [gps]*len(unique_trips), unique_trips))


def preprocess_trips(gps):
    gps = gps.reset_index(drop=True)
    gps = Polaris(gps) \
        .add_lag_of_column('velocity', lag_shifts=[1, 2, 3, 4, 5]) \
        .pandas_df_operation(func_name='fillna', columns=['velocity_lag_1', 'velocity_lag_2', 'velocity_lag_3',
                                                          'velocity_lag_4', 'velocity_lag_5'], arguments={'method': 'bfill'}) \
        .select_columns(['velocity', 'velocity_lag_1', 'velocity_lag_2', 'velocity_lag_3',
                         'velocity_lag_4', 'velocity_lag_5', 'is_speed_large']) \
        .df
    return gps

dataset = pd.concat(list(map(preprocess_trips, gps_trips)))
dataset

## X & y split

dataset.columns
X = dataset[['velocity', 'velocity_lag_1', 'velocity_lag_2', 'velocity_lag_3', 'velocity_lag_4', 'velocity_lag_5']]
y = dataset['is_speed_large']

## Train & test split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Modeling

from sklearn.linear_model import LogisticRegression

parameters = {
}

classifier = LogisticRegression(**parameters)
classifier.fit(X_train, y_train)


# Performance estimation

from sklearn.metrics import roc_auc_score


y_pred = classifier.predict(X_test)
HISTORY.append((parameters, roc_auc_score(y_test, y_pred)))

HISTORY

# MODEL EXPORT

import pickle


with open(OUTPUT_MODEL, 'wb') as file:
    pickle.dump(classifier, file)

classifier
