==> notebooks/PIPELINE_1_WASTE_COLLECTION_DETECTION/03_AUTOMATED_PIPELINE_TESTING.ipynb <==
# About

* **Author**: Adil Rashitov
* **Creation Date**: 22.02.2020
* **Goal**: This notebook is created to test automated version of pipeline(function) of GPSClustering. It is required to perform end-to-end optimization.
* **Deliverable**: Fast movement detecting pipeline

# Imports
import os
import numpy as np
import pandas as pd
import logging
import plotly.express as px
import geopandas as gpd
import plotly.express as px
from multiprocessing import Pool
from sklearn.model_selection import train_test_split
from GPSOdyssey import Polaris, Kepler, Void, Vega
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = 'all'

logger = logging.getLogger()
logger.setLevel(logging.INFO)

%load_ext autoreload
%autoreload 2

# Data

GPS_DATE = '2020-10-01'
S_MAP_MATCHING_REPORT = '/Data/Intermediate/MapMatchingReports/MapMatching_october2020.csv'

map_matching_report = pd.read_csv(S_MAP_MATCHING_REPORT, parse_dates=[
                                  'waste_collection_date']).sort_values(by=['waste_collection_date', 'truck_id'])

map_matching_report = map_matching_report[map_matching_report['waste_collection_date'] == GPS_DATE]
fnames = map_matching_report['csv_file']
fnames

## October GPS records

%%time
S_CLUSTERED_GPS = '/Data/Source/OctoberGPS/'


vega = Vega(engine='pandas')
gps = vega.read_from_dir(directory=S_CLUSTERED_GPS,
                             file_extensions='.csv',
                             args={'parse_dates': ['time'], 'dtype':{'lon': 'float', 'lat': 'float',
                                                                     'lon_match': 'float', 'lat_match': 'float',
                                                                    }},
                             concatenate=False, amt_in_parallel=6, filenames=fnames)
len(gps)

# GPS clustering pipeline

from sklearn.pipeline import Pipeline
from Pipeline import FastMovementPreprocessor, FastMovementClassifier, GPSDBSCAN


fastMovementPreprocessorArgs = {
    'select_columns': ['truck_id', 'lon', 'lat', 'lon_match', 'lat_match', 'time'],
    'datetime_column': 'time',
    'vehicle_id_col': 'truck_id',
    'lon_col': 'lon_match',
    'lat_col': 'lat_match'
}
fastMovementClassifierArgs = {
    'model_path': '/models/fast_movement_detector.sav',
    'X_cols': ['velocity', 'velocity_lag_1', 'velocity_lag_2',
               'velocity_lag_3', 'velocity_lag_4', 'velocity_lag_5'],
    'y_col': 'is_fast_moving'
}
args_GPSDBSCAN= {
    'eps': 130,
    'min_n_samples': 20,
    'cols2select': ['x', 'y', 'unixtime'],
    'fast_movement_col': 'is_fast_moving',
    'cluster_colname': 'cluster_id',
    'post_sort_cols': ['truck_id', 'unixtime'],
}


clustering_pipeline = Pipeline([
     ('feature_extractor', FastMovementPreprocessor(**fastMovementPreprocessorArgs)),
     ('FastMovementClassifier', FastMovementClassifier(**fastMovementClassifierArgs)),
     ('clustering_model', GPSDBSCAN(**args_GPSDBSCAN))
])

%%time
with Pool(4) as pool:
    gps_outputs = pool.starmap(clustering_pipeline.transform, zip(gps))

gps_outputs

# Kepler visualization

kepler = Kepler({"GPS": pd.concat(gps_outputs).reset_index(drop=True)}, height=800)
kepler.render_kepler_map()
kepler.get_rendered_map()
